requirements.txt
streamlit
streamlit-cropper
opencv-python-headless
Pillow
numpy
google.generativeai
imutils